requirements.txt
websockets>=12.0
anthropic>=0.40.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
